Adds typecast clause to convert all columns to text so we can search.
People have been reporting that whenever you included a non-text column
in the `searchable_columns` array, the functionality broke given it
couldn't perform a text ILIKE search on non these fields.

This commit fixes this, by doing a typecast on all fields declared
inside the `searchable_columns` array.

Also, we have appended `if` clauses to allow access to `JSON` views
in the browser for debugging purposes.